requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
dvc>=2.0.0
